+++ static_report/src/components/shared/Columns/ColumnCards/ColumnTypeDetail/ColumnTypeDetailText.tsx
@@ -4,12 +4,12 @@ import {
   formatColumnValueWith,
   formatIntervalMinMax,
 } from '../../../../../utils/formatters';
-import { getColumnMetricRatio } from '../../../../../utils/transformers';
 import { MetricCell } from '../../ColumnMetrics/MetricCell';
 import { DataCompositionMatrix } from '../../ColumnMatrices/DataCompositionMatrix';
 import { StatisticalMatrix } from '../../ColumnMatrices/StatisticalMatrix';
 import { UniquenessMetrics } from '../../ColumnMatrices/UniquenessMatrix';
 import { ZEROLENGTH } from '../../constants';
+import { getColumnMetricRatio } from '../../utils';
 
 type Props = { columnDatum: ColumnSchema };
 export const ColumnTypeDetailText: React.FC<Props> = ({ columnDatum }) => {
