+++ static_report/src/components/shared/Columns/ColumnMatrices/DataCompositionMatrix.tsx
@@ -5,9 +5,9 @@ import {
   formatColumnValueWith,
   formatIntervalMinMax,
 } from '../../../../utils/formatters';
-import { getColumnMetricRatio } from '../../../../utils/transformers';
 import { MetricCell } from '../ColumnMetrics/MetricCell';
 import { NULLS, INVALIDS } from '../constants';
+import { getColumnMetricRatio } from '../utils';
 
 type Props = { columnDatum: ColumnSchema; children?: ReactNode };
 export function DataCompositionMatrix({ columnDatum, children }: Props) {
