+++ static_report/src/components/shared/Columns/ColumnMatrices/UniquenessMatrix.tsx
@@ -4,9 +4,9 @@ import {
   formatColumnValueWith,
   formatIntervalMinMax,
 } from '../../../../utils/formatters';
-import { getColumnMetricRatio } from '../../../../utils/transformers';
 import { MetricCell } from '../ColumnMetrics/MetricCell';
 import { DISTINCTS, DUPLICATES } from '../constants';
+import { getColumnMetricRatio } from '../utils';
 
 type Props = { columnDatum: ColumnSchema };
 export function UniquenessMetrics({ columnDatum }: Props) {
