+++ static_report/src/components/shared/Columns/index.ts
@@ -1,6 +1,6 @@
 export * from './constants';
-export * from './ColumnTypeHeader';
-export * from './ColumnCards';
+// export * from './ColumnTypeHeader';
+// export * from './ColumnCards';
 export * from './ColumnDetailMasterList';
 export * from './ColumnMatrices';
 export * from './ColumnMetrics';
