+++ static_report/src/components/shared/Widgets/DataCompositionWidget.tsx
@@ -3,13 +3,13 @@ import { ColumnSchema } from '../../../sdlc/single-report-schema';
 import { Comparable } from '../../../types';
 import { renderChartUnavailableMsg } from '../Charts/utils';
 import { formatTitleCase } from '../../../utils/formatters';
-import {
-  containsAvgSDSummary,
-  transformCompositionAsFlatStackInput,
-} from '../../../utils/transformers';
 import FlatStackedBarChart from '../Charts/FlatStackedBarChart';
 import { GeneralStats } from '../Columns/ColumnMetrics/GeneralStats';
 import { TypedStats } from '../Columns/ColumnMetrics/TypedStats';
+import {
+  containsAvgSDSummary,
+  transformCompositionAsFlatStackInput,
+} from '../Columns/utils';
 
 interface Props extends Comparable {
   hasAnimation?: boolean;
