+++ static_report/src/lib/index.ts
@@ -3,6 +3,6 @@
  */
 export * from '../components/shared';
 // export * from '../utils';
-// export * from '../hooks';
+export * from '../hooks';
 // export * from '../types';
 // export * from '../sdlc';
